Separating various tracing elements
traceInfo is for internal dev. traceLogic is things that print to the
logic tracer for author to see. traceAlert are errors that possibly are
major issues; displays to viewer. e.g., loading answer file with
mismatched variable types or variable name issues. This is something
that could happen at user time so user should be informed since could
break an interview. They can complain to someone about it or ignore it.
See the SendFeedback button which is currently just spamming this. Since
it's possible for many errors, concatenate them to avoid many dialog
boxes. Will add a Report Error button.