Fix display of upload buttons.
There is an issue in CanJS where block elements can not be rendered
inside inline elements, except in anchor tags. Since #45 was merged, now
stache processes all the markup and in the process the content of the
upload buttons which was wrapped in `span` tag was being rendered
outside of its container. The quick fix to this was to change the `span`
containers to `a` tags.

Closes #48.