Go to clicked tab based on appState changes.
On page refresh if the hash routing was pointing to anything but
'templates', the app won't render the right page because the existing
code always defaults to the interviews list. This PR removes the click
binding of the navbar items on the legacy code and instead calls
`gotoTabOrPage` based on the appState change. This way the routing is
completely bi-directional, even on page refresh.

Closes #372.