Fix some issues in the drag & drop list
Calling `.attr` on the templates list in `updateSortOrder` caused the
model properties in the `define` object to be missing (they are not
'serialized') which caused the missing dates after sorting using drag &
drop. Since we were replacing the templates with new instances of the
model the logic to handle delete and restore was broken as well. This PR
fixes both issues by modifying the `sortOrder` directly in the bound
list, batching the events to avoid rendering the list in intermediate states.

Closes #424