Fix empty list messages display.
Some styles applied to the restored/deleted templates alert was messing
up the display of the messages shown when the templates list is empty,
or when there are no matches for the search results. This is fixed by
adding specific class names to the restored/deleted templates alert,
this way the other instances of the app-alert component on the page are
not affected.