Do not set `selected` to false in `element-container`
This property is bound to `editActive`, which should be set to false
only by the `a2j-template` that way, we can ensure the template saves
itself before the element leaves the "edit" mode. It also prevents some
weird issues where elements leave the "edit" mode when user is "editing"
its content. (e.g clicking the toolbar of a RTE).