Fix early error state in page fields.
`can-value` is causing `field._answer.values` to be set when the field
is rendered, this triggers the can.Map validation logic hance the error
states (inputs have no value at this point). The solution was to remove
the 2-way binding on `value` and update the inputs val on change using
the new binding syntax.

Closes #666.