Fix a bunch of issues related to old event syntax.
There were a bunch of event handlers not being working due to the usage
of the `(event)` syntax, which now is `($event)`. It also moves some
component dependencies to the template itself and fixes the root path of
CKEditor that stopped working after changing the root folder being
served by the static server.