Change template `component` property to `tag`.
Instead of holding a stache partial with the can.Component tag, it
holds the tagName that then it's converted in the client to its actual
can.Component representation. This is a first step to eliminate the
dynamic rendering of template nodes that obscures a little bit the way
elements (nodes) and its parent template (rootNode) communicate.