Initial implementation of the if-else element.
The content of the `if` and `else` blocks are hardcoded, this only
allows to duplicate/remove the existing element in that fixed template,
adding elements will be addressed in a different pull request. This PR,
adds the inital markup, and logic to define `if` conditions alongs with
the capability of evaluating said condition and rendering the correct
block (either `if` or `else`) during document assembly.