Remove circular reference in conditional element.
There was a circular reference caused by the way we were setting the
property `activeNode` in the `conditional-add-element` component. This
commit prevents `a2j-template` to set the `activeNode` to its child
elements. Instead `a2j-template` requires child elements that own nested
templates to implement a method that deselectes those childs when there
is a selection. Also, elements with nested templates are responsible of
allowing the selection of only one of its child elements at at time.

Closes #1028.