Remove intermediary event handler
Having a method in the viewModel with the same name as the property
received in the component is just a bad idea, might cause scoping
issues. In this case the callback passed to the component was called
instead of the one in the viewModel and since that callback was
expecting a parameter that was not provided in the template it failed.