Refactor interview form processing.
Before when the "Process Form" button was clicked, we'd call
`pState.save()` which was appending an html form to the body and
submitting it causing a redirect to the server's returned page. That
worked but the model was not the place to put that logic, instead we
moved it to the template and also added some code to handle the case
where we have to submit two forms, the asnwers and document assembly,
when the "Assemble and Save" button is clicked.