Fix A2JTemplate detection while looping to find child
The a2j-conditional component owns A2JTemplate instances where the
elements of the ifBody and elseBody are stored. When the parent
template loops through its children it needs to detect this nested
template instances in order to loop through the deep nested children
correctly. This detection logic has a bug causing the app to throw when
new instances of a2j-conditional were added to the document

Closes #1307