Fix Back to Prior Question navigation
Assigning a button in Author with the special value `Back to prior question`
was not implemented.  Set special behavior that overrides gotoPage and
button.next with the prior question name from the list of visited pages,
and prevents any After logic from executing to prevent unwanted side effects.
This does not take the user back to the prior question, but actually
navigates forward to it in order to preserve the complete page history.

Closes #1299