Fix A2J Dollar and DollarRound functions
Both Dollar and DollarRound were relying on the jQuery NumberFormatter
Plugin, which did not have an associated NPM package and was not currently
installed.  Found https://www.npmjs.com/package/numberformat
an NPM package based on the same jQuery plugin, but that did not
require jQuery as a dependency. Re-implemented the Dollar and
DollarRound A2J Functions using the `numberformat` package, as well as
the internal `niceNumber` function in tlogic.js, used to format large
Ordinal numbers.

Closes #1344