Fix author app buttons bindings
The file upload can.view.attr added recently imported A2J_AuthorApp
which is the file where most of the bindings are set, loading that file
before the app's html is in the DOM will cause event handlers to not be
attached correctly.

This commit changes the file so it doesn't load said file and replaces
the global CONST reference with the url for the guideupload "endpoint"
which is unlikely to change in the near future and also adds defensive
code to call the global "signin" function which should be available when
the file upload callback is executed.