Fix Steps starting on one instead of Zero
In the Steps tab while in A2J Author, the first step was being reset
to one instead of zero.  The problem was in the way A2J_Tabs.js was
preparing the DOM for view which was actually overwriting the current
Step values.  Also set MAXSTEPS to 13 as it was incorrectly set to
12.

Closes # 1426