Add Exit - User Does Not Qualify destination
Exit - UDNQ allows Authors to enter an external URL to help users
understand why they did not satisfy the requirements of the GI. It
only shows the URL field when this special destination is selected
for the current button.  Added Author messages for Exit options that
allow them to test their settings without actually leaving the interview.