Fix Audio not playing in popup pages
Pop up pages/modals were looking at the main page for the source of
their audio, video, and image files.  Modals now look to the pages
modalContent to get the correct source files. Refactored modal to be
more reusable and to use modelContent explicitly as it's data source.
Updated tests as well.

Closes #1227