Fix tracelogic error for TF variables
When using a TF variable, it was testing against a string value, but
if the value was already a boolean, it was throwing an error in the trace
logic.  Fxied the code to test for string type first. Technically, it should
always be a boolean value there, but as there many old interviews with loose
types, it is better to leave the string test.

Closes 1512