1538 - Add standard Fail message modal if no URL
If an Author forgets or does not have a URL with more info to show
when a user fails an interview, it now shows a Fail message in
a modal.  Code now checks the author entered Fail url for a proper
protocol `http` or `https`, and if it's not there, adds `http` to allow
the url to load in a new tab or window. Refactored alerts in Author to
use modals now as well.

Closes 1538