Fix internal list built from options list
Internal lists are now captured at the time the <options> list is
generated and stored on Field.previousTextList, for later retrieval.
If field.previousTextList is empty string or undefined (uploaded GIs)
it will rebuild the text list with hard returns from the currently
saved <options> list html string in Field.listData as before.

Closes 300