Implement calculator for number and numberdollar
Added jquery plugin to allow calculator for number and numberdollar
fields only when selected by the Author. Selecting 'show calculator'
renders a small calculator icon underneath the field which launches
the jquery plugin calculator when clicked. The onClose callback fires
the validateField function on the calculated value to make sure it gets
saved to the answer file.

Closes #212