Fix default values being set to fields
Fields with default values now get set if there is no previous answer
saved for that field. Users will see those default values on page load
and if those are not changed, they will be saved to the answer set as
though the user had entered or selected that answer themselves. Changed
or previous answers will override default values as expected.

Closes 1590