1617 Update page name when edit dialog closes
Changing a page name in Author was not being updated the same way it
was when hitting the Close button. Updates now fire on the close event
which works regardless of how the edit page dialog closes. Guide is now
saved when the dialog closes as well, which persists changes if an Author
should leave the web page either on purpose or accident (which happend to
me).

Closes 1617