1637 Fix file name issues
Unsafe window filename characters / ? < > \ : * | " found in GI
titles are removed from filenames when downloading A2J interviews.
Forward slashes are replace with hyphens to preserve dates.
ie: 5/27/15 becomes 5-27-15. Titles with apostrophes in them are
escaped by mysql when being saved to the database, and were being
displayed in the Interviews tab as such. Those back slashes are now
removed for display. Original titles in the About tab of an interview
are not affected, and will remain with as saved by the Author.

Closes 1637