1658 Fix answer clearing in Author preview models
Long var/answer lists were taking a long time to clear and freezing
up the browser. The `each` loop now only resets answers that have a
`values` property which have stored answers - aka that are not the
default array of [null]. During testing, notice that a file upload if
abandoned throws an `undefined` error in the console, and updated the
check to make sure `file` is defined.

Closes 1658