1396 Fix authors tab info not saving
The Authors sub-tab under the About tab was clearing out any previously
saved author lists. The code used to populate the author list assumed that
the table would not be hidden at the time it was being populated, the same
as the Steps tab/list. However, due to the Authors tab being a hidden sub tab
of the about tab, it was hidden, and therefore creating an empty list, saving
over and wiping out any previously loaded data. Now it detects if the list loading
is the Authors list, and uses the previously saved list on first load, fixing the
issue. A current side effect is that previously saved GIs with an empty list will
still show a blank selector until they are saved the first time with Author information.

Closes 1396