Fix required checkbox on NOTA and Radio Buttons
CheckboxNOTA (none of thee above) and Radio Buttons are set to required
by default, and hide the required checkbox from Authors. Because the box
is never actually checked, but just hidden, if an Author selects a different
field type, the checkbox for required would show unchecked even though the
field itself was still required in the code. This places the checkmark in the
required box before it is hidden.

Closes 1672