1805 citation fields (#1839)
* citations set and saving

* full report updated for citations

* re-map styling to css file for reports

* full report stache creation basics

* language and variable formatting working

* parity display

* page iteration and filtering

* text statistics npm package added

* standardize and cleanup styling

* jump links working

* updated toolbar

* escaped html tags from text fields

* print button working and print styling

* performance testing

* performance updates for Edge

* popups seperated and styling fixes

* report cleanups from jessica feedback

* Handle building of first full report

Large GIs were causing the report tab to not render anything, and make
appear the app was stuck on the current tab. Handled this with a
combination of using a promise to resolve the completion of the report
and a 'building report' spinner. Also disables print button when no
guide is loaded

* print media styles set

* fix regex syntax error for jshint

* fixed NaN error for page fields

* added live documenation to report and report toolbar

* tests updated and passing

* fixed jshint error in tests

This satisfies the grant requirement legal citations in interviews and a report to review them. It also refactors the report tab to a full CanJS component with documentation and tests.