update ie11 validation handler to include date types (#1870)
* handles ie11 dates when tabbing through fields

* Handle IE11 focus event bug for dates

Tabbing through fields with the jquery-ui datepicker does not always
open the datepicker calendar, preventing the close event which validates
the selected date onClose. Rewired the close event to use a separate
function which is also called in IE11 to check dates for validation when
entering them directly via text values.

Closes 1831

* handles ie11 dates when tabbing through fields

* Handle IE11 focus event bug for dates

Tabbing through fields with the jquery-ui datepicker does not always
open the datepicker calendar, preventing the close event which validates
the selected date onClose. Rewired the close event to use a separate
function which is also called in IE11 to check dates for validation when
entering them directly via text values.

Closes 1831