1958 fix cmp file upload not refreshing variables (#1959)
* refactored cmp upload working with canjs

* fix syntax for jshint

* Update appState change event for vars

The appState was binding on any change at the appState level and testing
the %root attr to see if the change was on `guide.vars`. This causes a
cascade of bubbling events severly affectiong performance across the app
and specifically during cmp uploads. Justin helped me write a custom
compute to listening only on `guide.vars` changes directly, which is also
able to wait until the batch process of data from a cmp upload has
completed before re-rendering the view.