Fix bug editing an element within a conditional while editing the conditional
To reproduce:
1. Edit an a2j-conditional (double click, a popup will display)
2. Edit an element within that conditional (double click on the element)

The issue: the a2j-conditional edit popup would still be displayed.

Cause: conditionals have an a2j-template within them, so their `rootNode' was
/that/ template, and /not/ the outermost page template, so the conditional's
`editActive' wasn't being changed.

The odd thing about this is that it works before the upgrade, where (I'd assume)
we /should/ have the same confusion with `this.attr("rootNode")' within a
conditional. I'm not sure why it works in prod.